retrofit implemented. List is displayed

+++ app/src/main/java/com/example/revoluttest/model/Currency.kt
@@ -1,6 +1,8 @@
 package com.example.revoluttest.model
 
-class Currency(val currencyName : String,
-               val countryCurrencyName: String,
-               val flag: Int,
-               val value: Float)
+class Currency(
+    val currencyName: String,
+    val countryCurrencyName: String,
+    val flag: Int,
+    val value: Double
+)
